Duplicate File

+++ documents_server_folder/models/documents_document.py
@@ -253,7 +253,7 @@ class Document(models.Model):
             names.append(doc.name)
         return "/".join(names)
 
-    def get_full_path(self):
+    def get_full_path(self, check_exist=True):
         self.ensure_one()
         if not self.located_on_the_server:
             return None
@@ -263,8 +263,10 @@ class Document(models.Model):
         if not is_abs:
             raise FileNotFoundError("File not found: " + full_path)
         full_path = os.path.normpath(os.path.normcase(full_path))
-        if not os.path.exists(full_path) and self.active:
-            raise FileNotFoundError("File not found: " + full_path)
+        if self.active and not os.path.exists(full_path):
+            if check_exist:
+                raise FileNotFoundError("File not found: " + full_path)
+            return None
         return full_path
 
     @api.model_create_multi
@@ -290,21 +292,32 @@ class Document(models.Model):
             old_path = self.get_full_path()
             folder_path = os.path.dirname(old_path)
             new_path = os.path.join(folder_path, name)
-            os.rename(old_path, new_path)
+            if old_path != new_path:
+                os.rename(old_path, new_path)
         return super().write(vals)
 
     def toggle_active(self):
         server_items = self.filtered("located_on_the_server")
         active_items = server_items.filtered(self._active_name)
         for item in active_items:
-            path = item.get_full_path()
-            if os.path.isdir(path):
-                shutil.rmtree(path)
-            else:
-                os.remove(path)
+            path = item.get_full_path(check_exist=False)
+            if path:
+                if os.path.isdir(path):
+                    shutil.rmtree(path)
+                else:
+                    os.remove(path)
         inactive_items = self - active_items
         folders_to_restore = inactive_items.filtered(lambda x: x.type == "folder")
         for folder in folders_to_restore:
             path = folder.get_full_path()
             os.makedirs(path)
         super().toggle_active()
+
+    def copy(self, default=None):
+        server_items = self.filtered("located_on_the_server")
+        server_files = server_items.filtered(lambda x: x.type != "folder")
+        for file in server_files:
+            old_path = file.get_full_path()
+            new_path = old_path + " (copy)"
+            shutil.copyfile(old_path, new_path)
+        return super().copy(default)
